requirements.txt
Django
djangorestframework
Pillow
djangorestframework-jwt
psycopg2
